refactor: organize raycaster files
